requirements.txt
 langchain
 python-dotenv